Pass delete flag when calling git branch with args
Previously, trying to call gbr -d to delete a branch was resulting in an
error message stating that the branch already existed being printed
since the -d flag was not passed to the git branch command. This commit
fixes this error by correctly passing the flag if and only if it is set.